BJS2-66383 minor fix
